move run_volumes and exclusions to pipeline setup step; use SQLite database for l1_confound storage and lookup

+++ R/insert_df_sqlite.R
@@ -4,6 +4,12 @@
 #' @param sesssion the session of these data
 #' @param run_number the run_number of these data
 #' @param data A \code{data.frame} containing the data to be inserted into the sqlite db
+#' @param table A character string of the table name to be modified
+#' @param delete_extant Whether to delete any existing records for this id + session + run_number combination
+#' @param append Whether to append records to the table (passed through to dbWriteTable)
+#' @param overwrite Whether to overwrite the existing table (passed through to dbWriteTable)
+#' 
+#' @return a TRUE/FALSE indicating whether the record was successfully inserted
 #' @importFrom checkmate assert_integerish test_null assert_data_frame assert_string
 insert_df_sqlite <- function(gpa = NULL, id = NULL, session = NULL, run_number = NULL, data = NULL,
                              table = NULL, delete_extant = TRUE, append = TRUE, overwrite = FALSE) {
@@ -26,7 +32,7 @@ insert_df_sqlite <- function(gpa = NULL, id = NULL, session = NULL, run_number =
   }
 
   # delete any existing record
-  if (isTRUE(delete_extant)) {
+  if (isTRUE(delete_extant) && DBI::dbExistsTable(con, table)) {
     query <- glue::glue_sql(
       "DELETE FROM {`table`}",
       "WHERE id = {id} AND session = {session}",
@@ -36,7 +42,7 @@ insert_df_sqlite <- function(gpa = NULL, id = NULL, session = NULL, run_number =
     DBI::dbExecute(con, query)
   }
 
-  # add record
+  # add record -- include keying fields for lookup
   data$id <- id
   data$session <- session
   if (!is.null(run_number)) data$run_number <- run_number
@@ -48,8 +54,6 @@ insert_df_sqlite <- function(gpa = NULL, id = NULL, session = NULL, run_number =
       return(FALSE)
     }
   )
-
-  DBI::dbReadTable(con, "test")
 
   return(res)
 }
